requirements.txt
langchain-core
langchain-community
langchain-groq
pandas
streamlit
python-dotenv


